requirements.txt
tqdm
rotary_embedding_torch
transformers==4.19
tokenizers
inflect
progressbar
einops==0.4.1
unidecode
scipy==1.10.1
librosa==0.9.1
ffmpeg
numpy==1.20.3
torchaudio
threadpoolctl
appdirs